- выглаживаю "безопасность". Более точное соответствие ролям "Добавить, удалить, редактировать, смотреть"..

+++ controller/server/places/libre_place_sub.php
@@ -36,8 +36,8 @@ if (isset($_POST["comment"]))      {$comment= $_POST['comment'];}    else {$comm
 
 
 //echo "!$placesid!";
-if ($oper=='')
-{
+if ($oper==''){
+    if (($user->mode==1) or ($user->TestRoles('1,3'))){    
 	if(!$sidx) $sidx =1;
 	$result = $sqlcn->ExecuteSQL("SELECT COUNT(*) AS count FROM places_users WHERE placesid='$placesid'");
 	$row = mysqli_fetch_array($result);
@@ -63,23 +63,27 @@ if ($oper=='')
 	    $i++;
 	}
 	echo json_encode($responce);
+    };
 };
-if ($oper=='edit')
-{
+if ($oper=='edit'){
+    if (($user->mode==1) or ($user->TestRoles('1,5'))){    
 	$SQL = "UPDATE places_users SET userid='$name' WHERE id='$id'";
 	$result = $sqlcn->ExecuteSQL( $SQL ) or die("Не могу обновить данные по помещениям/пользователям!".mysqli_error($sqlcn->idsqlconnection));
+    };
 };
-if ($oper=='add')
-{
+if ($oper=='add'){
+    if (($user->mode==1) or ($user->TestRoles('1,4'))){        
         if (($placesid=="") or ($name=="")) {die();};
 	$SQL = "INSERT INTO places_users (id,placesid,userid) VALUES (null,'$placesid','$name')";
 	$result = $sqlcn->ExecuteSQL( $SQL ) or die("Не могу добавить помещение/пользователя!".mysqli_error($sqlcn->idsqlconnection));
+    };
 
 };
-if ($oper=='del')
-{
+if ($oper=='del'){
+    if (($user->mode==1) or ($user->TestRoles('1,6'))){    
 	$SQL = "DELETE FROM places_users WHERE id='$id'";
 	$result = $sqlcn->ExecuteSQL( $SQL ) or die("Не могу удалить помещение/пользователя!".mysqli_error($sqlcn->idsqlconnection));
+    };
 };
 
 ?>
